KS-2-2 fixed payroll calc

+++ src/main/kotlin/lesson_2/lesson_2_2.kt
@@ -11,7 +11,7 @@ fun main() {
     val payrollAllStudents = countStudents * payrollStudent
     val allPayroll = payrollAllWorkers + payrollAllStudents
 
-    val averagePayroll = (allPayroll / (countWorkers + countStudents)).toInt()
+    val averagePayroll = allPayroll / (countWorkers + countStudents)
 
     println(payrollAllWorkers)
     println(allPayroll)
